[views] Export namespaces to be used for app classes and globals
Instead of using the global window object as the place for all things
global, we better define more intuitive and readable namespaces to be
used throughout the app client-side code. The 'PaperBox' namespace is
to be used for app-specific classes and 'Global' will hold all global
variables (and should be used with care).