[entries] Improve smooth scrolling animation handling
When next/previous key is used while the animation is running, we'd key
inconsistent results depending on timing with active entry changing
almost randomly. In order to fix that, all running animations are
stopped before starting a new one and the usual scroll handlers are
disable while animating and only re-enabled once the animation is over.